requirements.txt
-e .
pytest==6.2.4
aws_cdk.assertions==1.112.0
